Working implementation of cloning examples organization on user sign up
