+++ app/server/appsmith-server/src/main/java/com/appsmith/server/repositories/ApplicationRepository.java
@@ -2,8 +2,11 @@ package com.appsmith.server.repositories;
 
 import com.appsmith.server.domains.Application;
 import org.springframework.stereotype.Repository;
+import reactor.core.publisher.Flux;
 
 @Repository
 public interface ApplicationRepository extends BaseRepository<Application, String>, CustomApplicationRepository {
+
+    Flux<Application> findByOrganizationId(String organizationId);
 
 }
