+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationService.java
@@ -3,7 +3,6 @@ package com.appsmith.server.services;
 import com.appsmith.server.acl.AclPermission;
 import com.appsmith.server.domains.Application;
 import com.appsmith.server.dtos.ApplicationAccessDTO;
-import com.appsmith.server.dtos.UserHomepageDTO;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -24,8 +23,6 @@ public interface ApplicationService extends CrudService<Application, String> {
     Mono<Application> save(Application application);
 
     Mono<Application> archive(Application application);
-
-    Mono<UserHomepageDTO> getAllApplications();
 
     Mono<Application> changeViewAccess (String id, ApplicationAccessDTO applicationAccessDTO);
 }
