+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationServiceImpl.java
@@ -8,12 +8,9 @@ import com.appsmith.server.domains.Action;
 import com.appsmith.server.domains.Application;
 import com.appsmith.server.domains.ApplicationPage;
 import com.appsmith.server.domains.Layout;
-import com.appsmith.server.domains.Organization;
 import com.appsmith.server.domains.Page;
 import com.appsmith.server.domains.User;
 import com.appsmith.server.dtos.ApplicationAccessDTO;
-import com.appsmith.server.dtos.OrganizationApplicationsDTO;
-import com.appsmith.server.dtos.UserHomepageDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
 import com.appsmith.server.helpers.PolicyUtils;
@@ -31,17 +28,12 @@ import reactor.core.scheduler.Scheduler;
 
 import javax.validation.Validator;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import static com.appsmith.server.acl.AclPermission.MANAGE_APPLICATIONS;
 import static com.appsmith.server.acl.AclPermission.READ_APPLICATIONS;
-import static com.appsmith.server.acl.AclPermission.READ_ORGANIZATIONS;
 
 
 @Slf4j
@@ -51,9 +43,6 @@ public class ApplicationServiceImpl extends BaseService<ApplicationRepository, A
     //Using PageRepository instead of PageService is because a cyclic dependency is introduced if PageService is used here.
     //TODO : Solve for this across LayoutService, PageService and ApplicationService.
     private final PageRepository pageRepository;
-    private final SessionUserService sessionUserService;
-    private final OrganizationService organizationService;
-    private final UserService userService;
     private final PolicyUtils policyUtils;
 
     @Autowired
@@ -64,15 +53,9 @@ public class ApplicationServiceImpl extends BaseService<ApplicationRepository, A
                                   ApplicationRepository repository,
                                   AnalyticsService analyticsService,
                                   PageRepository pageRepository,
-                                  SessionUserService sessionUserService,
-                                  OrganizationService organizationService,
-                                  UserService userService,
                                   PolicyUtils policyUtils) {
         super(scheduler, validator, mongoConverter, reactiveMongoTemplate, repository, analyticsService);
         this.pageRepository = pageRepository;
-        this.sessionUserService = sessionUserService;
-        this.organizationService = organizationService;
-        this.userService = userService;
         this.policyUtils = policyUtils;
     }
 
@@ -172,78 +155,6 @@ public class ApplicationServiceImpl extends BaseService<ApplicationRepository, A
                         .flatMap(pageRepository::save))
                 .collectList()
                 .map(pages -> true);
-    }
-
-    /**
-     * For the current user, it first fetches all the organizations that its part of. For each organization, in turn all
-     * the applications are fetched. These applications are then returned grouped by Organizations in a special DTO and returned
-     *
-     * TODO : We should also return applications shared with the user as part of this.
-     *
-     * @return List of OrganizationApplicationsDTO
-     */
-    @Override
-    public Mono<UserHomepageDTO> getAllApplications() {
-
-        Mono<User> userMono = sessionUserService
-                .getCurrentUser()
-                .flatMap(user -> {
-                    if (user.getIsAnonymous()) {
-                        return Mono.error(new AppsmithException(AppsmithError.USER_NOT_SIGNED_IN));
-                    }
-                    return Mono.just(user.getUsername());
-                })
-                .flatMap(userService::findByEmail)
-                .cache();
-
-        return userMono
-                .flatMap(user -> {
-                    Set<String> orgIds = user.getOrganizationIds();
-                    /*
-                     * For all the organization ids present in the user object, fetch all the organization objects
-                     * and store in a map for fast access;
-                     */
-                    Mono<Map<String, Organization>> organizationsMapMono = organizationService.findByIdsIn(orgIds, READ_ORGANIZATIONS)
-                            .collectMap(Organization::getId, Function.identity());
-
-                    UserHomepageDTO userHomepageDTO = new UserHomepageDTO();
-                    userHomepageDTO.setUser(user);
-
-                    return repository
-                            // Fetch all the applications which belong the organization ids present in the user
-                            .findByMultipleOrganizationIds(orgIds, READ_APPLICATIONS)
-                            // Collect all the applications as a map with organization id as a key
-                            .collectMultimap(Application::getOrganizationId)
-                            .zipWith(organizationsMapMono)
-                            .map(tuple -> {
-                                Map<String, Collection<Application>> applicationsCollectionByOrgId = tuple.getT1();
-                                Map<String, Organization> organizationsMap = tuple.getT2();
-
-                                List<OrganizationApplicationsDTO> organizationApplicationsDTOS = new ArrayList<>();
-
-                                Iterator<Map.Entry<String, Organization>> orgIterator = organizationsMap.entrySet().iterator();
-
-                                while (orgIterator.hasNext()) {
-                                    Map.Entry<String, Organization> organizationEntry = orgIterator.next();
-                                    String orgId = organizationEntry.getKey();
-                                    Organization organization = organizationEntry.getValue();
-                                    Collection<Application> applicationCollection = applicationsCollectionByOrgId.get(orgId);
-
-                                    List<Application> applicationList = new ArrayList<>();
-                                    if (applicationCollection!=null && !applicationCollection.isEmpty()) {
-                                        applicationList = applicationCollection.stream().collect(Collectors.toList());
-                                    }
-
-                                    OrganizationApplicationsDTO organizationApplicationsDTO = new OrganizationApplicationsDTO();
-                                    organizationApplicationsDTO.setOrganization(organization);
-                                    organizationApplicationsDTO.setApplications(applicationList);
-
-                                    organizationApplicationsDTOS.add(organizationApplicationsDTO);
-                                }
-                                userHomepageDTO.setOrganizationApplications(organizationApplicationsDTOS);
-                                return userHomepageDTO;
-                            });
-                });
     }
 
     @Override
